requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pandas==2.2.2
openpyxl==3.1.5
python-decouple==3.8
django-cors-headers==4.5.0

